% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConstructor.R
\docType{class}
\name{classconstructor}
\alias{classconstructor}
\alias{dsws}
\title{dsws}
\description{
An R5/RC object for accessing the LSEG Datastream
  DSWS service.
}
\details{
Creates an R5/RC4 object for accessing the LSEG
  Datastream DSWS service
}
\section{Fields}{

\describe{
\item{\code{tokenList}}{fieldDescription}

\item{\code{tokenSource}}{fieldDescription}

\item{\code{serverURL}}{fieldDescription}

\item{\code{username}}{fieldDescription}

\item{\code{password}}{fieldDescription}

\item{\code{initialised}}{fieldDescription}

\item{\code{errorlist}}{fieldDescription}

\item{\code{requestList}}{fieldDescription}

\item{\code{jsonResponseSaveFile}}{fieldDescription}

\item{\code{jsonResponseLoadFile}}{fieldDescription}

\item{\code{dataResponse}}{fieldDescription}

\item{\code{symbolList}}{fieldDescription}

\item{\code{myValues}}{fieldDescription}

\item{\code{myTypes}}{fieldDescription}

\item{\code{logging}}{fieldDescription}

\item{\code{numDatatype}}{fieldDescription}

\item{\code{numInstrument}}{fieldDescription}

\item{\code{numRequests}}{fieldDescription}

\item{\code{numChunks}}{fieldDescription}

\item{\code{chunkLimit}}{fieldDescription}

\item{\code{requestStringLimit}}{fieldDescription}

\item{\code{logFileFolder}}{fieldDescription}
}}

\section{Methods}{

\describe{
\item{\code{UCTSUpload(
  tsData,
  TSCode = "",
  MGMTGroup = "MGMTG",
  freq = c("D", "W", "M", "Q", "Y"),
  seriesName,
  Units = "",
  Decimals = 2,
  ActPer = c("N", "Y"),
  freqConversion = c("ACT", "SUM", "AVG", "END"),
  Alignment = c("1ST", "MID", "END"),
  Carry = c("YES", "NO", "PAD"),
  PrimeCurr = ""
)}}{Uploads an xts into a UCTS in the Datastream Database

Note this function does not check to see if there is
a pre-existing timeseries already in Datastream.  It will just overwrite
any existing UCTS.

Parameters are: \describe{
  \item{tsData}{an xts (or timeseries object that can be converted to one) to be uploaded.}
  \item{TSCode}{The mnemonic of the target UCTS}
  \item{MGMTGroup}{Must have managment group.  Only the first characters will be used.}
  \item{freq}{The frequency of the data to be uploaded}
  \item{seriesName}{the name of the series - can be no more than XX characters -  excess will be trimmed to that length}
  \item{Units}{Units of the data - can be no more than 12 characters - excess will be trimmed to that length}
  \item{Decimals}{Number of Decimals in the data - a number between 0 and 9 - if outside that range then trimmed}
  \item{ActPer}{Whether the values are percentages ("N") or actual numbers ("Y")}
  \item{freqConversion}{How to do any FX conversions}
  \item{Alignment}{Alignment of the data within periods}
  \item{Carry}{whether to carry data over missing dates}
  \item{PrimeCurr}{the currency of the timeseries}
}
Returns TRUE if the upload has been a success, or FALSE with field `errorlist` containing the error message

  Examples:

    \preformatted{
    mydsws <- dsws$new()

    sPost <- mydsws$UCTSUpload(tsData = testData,
                               TSCode = "TSTEST01",
                               MGMTGroup = "TEST",
                               freq = "D",
                               seriesName = "UCTS Upload Test",
                               Units = "",
                               Decimals = 2,
                               ActPer = "Y",
                               freqConversion = "END",
                               Alignment = "MID",
                               Carry = "NO",
                               PrimeCurr = "")

    expect_equal(sPost, TRUE)
    expect_equal(mydsws$errorlist, NULL)
}
  }

\item{\code{initialize(
  dsws.serverURL = "",
  getTokenFunction = NULL,
  token = NULL,
  username = "",
  password = "",
  connect = TRUE
)}}{  initialises the class.
  Unless noConnect is TRUE also connects to the
  Datastream dsws server.

  Authentication can be set in three ways:
  1)  If getTokenFunction is not null then that function is called.  It is expected to
  return a list with items 'TokenValue' and 'TokenExpiry'.

  2)  An access token can also be passed into the class on initialisation, so that it can be shared between sessions.
  'token' is expected to be a list with items 'TokenValue' and 'TokenExpiry'.

  3) A username and password that are used to fetch a token from the DSWS server.  If the username and password are not
  provided, then they are sourced from system enviroment variables (ie Sys.getenv)
      'DatastreamUsername' and 'DatastreamPassword'
  or alternatively (not preferred) then from
          \code{options()$Datastream.Username} and
          \code{options()$Datastream.Password}

    This allows the password to be stored in .Renviron or .RProfile rather
  than in the source code.

  There different accounts have different limits according to their licence.  Most users are limited to 50 items while
  enterprise users have a limit of 2000L.  The chunk limit
  can be controlled by setting the chunkLimit parameter of the dsws object.  If     \code{options()$Datastream.ChunkLimit} is
  set then the value is taken from there.

  }

\item{\code{listRequest(instrument, datatype = "", expression = "", requestDate)}}{  Make a listRequest from Datastream DSWS.
  This is the equivalent to
  the Excel static request for a list.

  Parameters are: \describe{
      \item{instrument}{should contain a list mnemonic, such as 'LFTSE100'
      Can be a user created list or index.  The UCL can contain
      expressions}
      \item{datatype}{array of datatypes eg NAME, MNEM, P, PE etc}
      \item{expression}{if datatype is null or '' then an expression
      eg PCH#(XXXX,3M)}
      \item{requestDate}{either a Date or a string with a datastream
      relative date eg '-3M'}
  }
  Returns a data.frame with the requested data.


  Examples:

    \preformatted{
      mydsws$listRequest(instrument = "LFTSE100",
      datatype = c("NAME","P"), 
requestDate = "-0D")
    }
    \preformatted{
      mydsws$listRequest(instrument = "LFTSE100",
      expression = "PCH#(XXXX,3M)", requestDate = Sys.Date())
    }
  }

\item{\code{snapshotRequest(instrument, datatype = "", expression = "", requestDate)}}{  Make a snapshotRequest from Datastream DSWS.
  This is the equivalent
  to the Excel static request for an array of instruments.

  Parameters are: \describe{
  \item{instrument}{should one or more instruments eg "MKS" or
      c("MKS","@AAPL").  The array can contain
      Economics codes and Expressions. }
  \item{datatype}{array of datatypes eg NAME, MNEM, P, PE etc}
  \item{expression}{if datatype is null or '' then an expression
      eg PCH#(XXXX,3M)}
  \item{requestDate}{either a Date or a string with a datastream relative
      date eg '-3M'}
}
  Returns a data.frame with the requested data.


  Examples:

    \preformatted{
  mydsws$snapshotRequest(instrument = c("MKS","@AAPL"), 

   datatype = c("NAME","P"), requestDate = "-0D")
    }
    \preformatted{
  mydsws$snapshotRequest(instrument = c("MKS","@AAPL"), 

   expression = "PCH#(XXXX,3M)", requestDate = "-0D")
    }

  }

\item{\code{timeSeriesListRequest(
  instrument,
  datatype = "",
  expression = "",
  startDate,
  endDate,
  frequency = "D",
  format = "ByInstrument"
)}}{  Make a timeSeriesListRequest from Datastream DSWS.
  This is the
  equivalent to the Excel timeseries request for an array of instruments.
  Should request either a datatype or an expression not both.  If a
  datatype is provided then anything in Expression will be ignored.

  Parameters are: \describe{
  \item{instrument}{should contain a list mnemonic, such as "LFTSE100"
.
  Can be a user created list or index.  The UCL can contain expressions.}
  \item{datatype}{array of datatypes eg P, PE etc}
  \item{expression}{if datatype is null or '' then an expression 

    eg PCH#(XXXX,3M)}
  \item{startDate}{either a Date or a string with a datastream relative date

    eg '-3M'}
  \item{endDate}{either a Date or a string with a datastream relative date 

    eg '-0D'}
  \item{frequency}{one of the standard Datastream frequencies - D, W, M, Q, or Y }
  \item{format}{can be either  "ByInstrument" or "ByDatatype". }
  }
  Returns either a single xts or a list of xts a data.frame with
  the requested data.  If "ByInstrument" then the data is returned as
  one or more (ie a list) wide xts with one column per instrument.
  If "ByDatatype" then the data is returned as one or more (ie a list)
  of wide xts with one column per Datatype.  This format is more compatible
  with the quantmod package.

  Examples:

    \preformatted{
  mydsws$timeSeriesListRequest(instrument = "LFTSE100",

    datatype = "P", startDate = "-30D",

    endDate = "-0D", frequency = "D")
    }
    \preformatted{

  mydsws$timeSeriesListRequest(instrument = "LFTSE100", 

    expression = "PCH#(XXXX,3M)", 

    startDate = "-30D",

    endDate = "-0D", 

    frequency = "D")
    }
    \preformatted{
  mydsws$timeSeriesListRequest(instrument = "LFTSE100", 

    datatype = ("P","UP"), startDate = "-30D",

    endDate = "-0D", 

    frequency = "D", format = "ByDatatype")
}
  }

\item{\code{timeSeriesRequest(
  instrument,
  datatype = "",
  expression = "",
  startDate,
  endDate,
  frequency = "D",
  format = "ByInstrument"
)}}{  Return a timeSeriesRequest from Datastream dsws.
  Should request either
  a datatype or an expression
  not both.  If a datatype is provided then anythink in Expression
  will be ignored.

  Make a timeSeriesRequest from Datastream DSWS.  This is the equivalent
  to the Excel timeseries request for an array of instruments.

  Parameters are: \describe{
  \item{instrument}{should one or more instruments eg "MKS" or
   c("MKS","@AAPL").  The array can contain Economics codes and
     Expressions. }
  \item{datatype}{array of datatypes eg P, PE etc}
  \item{expression}{if datatype is null or '' then an expression
  eg PCH#(XXXX,3M)}
  \item{startDate}{either a Date or a string with a datastream relative
  date eg '-3M'}
  \item{endDate}{either a Date or a string with a datastream relative
  date eg '-0D'}
  \item{frequency}{one of the standard Datastream
    frequencies - D, W, M, Q, or Y}
  \item{format}{can be either  "ByInstrument" or "ByDatatype".}
}
  Returns either a single xts or a list of xts a data.frame with
  the requested data.  If "ByInstrument" then
  the data is returned as one or more (ie a list) wide xts with one
  column per instrument.  If "ByDatatype"
  then the data is returned as one or more (ie a list) of wide xts with
  one column per Datatype.  This format
  is more compatible with the quantmod package.

  Examples:

    \preformatted{
  mydsws$timeSeriesRequest(instrument = c("MKS","@AAPL"),

      datatype = "P", startDate = "-30D",

      endDate = "-0D", frequency = "D")
    }
    \preformatted{

  mydsws$timeSeriesRequest(instrument = c("MKS"), 

      expression = "PCH#(XXXX,3M)", startDate = "-30D",

      endDate = "-0D", frequency = "D")
    }
    \preformatted{
  mydsws$timeSeriesRequest(instrument = c("MKS","@AAPL"), 

      datatype = ("P","UP"), startDate = "-30D",

      endDate = "-0D", frequency = "D", format = "ByDatatype")
    }

  }
}}

\examples{
\dontrun{


     mydsws <- dsws$new()
     # Snapshot requests


     myData <- mydsws$snapshotRequest(instrument = c("ABF","RIO","WPP"),
                                      datatype = "P",
                                      requestDate = "0D")

     myData <- mydsws$snapshotRequest(instrument = c("ABF","RIO","WPP"),
                                      expression = "PCH#(XXXX,3M)",
                                      requestDate = "0D")
     myData <- mydsws$listRequest(instrument = "LFTSE100", datatype = "P", requestDate = "0D")

     mydsws$snapshotRequest(instrument = c("SWCNB10","UKEUSCCIR"),
                            datatype = c("MNEM","UPDATE"),
                            requestDate = "0D")
     mydsws$snapshotRequest(instrument = c("VOD", "HSBA"),
                            datatype="QTEALL",
                            requestDate = Sys.Date())
     mydsws$snapshotRequest(instrument = "STATS",
                            datatype = "DS.USERSTATS",
                            requestDate = Sys.Date())


     # Timeseries requests


     xtsData <- mydsws$timeSeriesRequest(instrument = "MKS",
                                         datatype = "MV",
                                         startDate = "-30D",
                                         endDate = "-0D",
                                         frequency = "D")

     xtsData <- mydsws$timeSeriesListRequest(instrument = "LFTSE100",
                                             datatype = "MV",
                                             startDate = "-30D",
                                             endDate = "-0D",
                                             frequency = "D")

     # UCTS upload

       testData <- xts::xts(x = c(1, 2.2, 3.12345, 4.5),
       order.by = as.Date(c("2014-04-22","2014-04-23","2014-04-24","2014-04-25")))

       mydsws <- dsws$new()

       sPost <- mydsws$UCTSUpload(TSCode = "TSTEST01",
                                  MGMTGroup = "TEST",
                                  freq = "D",
                                  seriesName = "Upload Demonstration",
                                  Units = "",
                                  Decimals = 2,
                                  ActPer = "Y",
                                  freqConversion = "END",
                                  Alignment = "MID",
                                  Carry = "NO",
                                  PrimeCurr = "",
                                  tsData = testData)

}
}
