% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatastreamDSWS2R.R
\docType{package}
\name{DatastreamDSWS2R-package}
\alias{DatastreamDSWS2R}
\alias{DatastreamDSWS2R-package}
\title{@docType package
  @name DatastreamDSWS2R}
\description{
A package to manage access to the Thomson Reuters Datastream DSWS webservice
}

