% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatastreamDSWS2R.R
\docType{package}
\name{DatastreamDSWS2R}
\alias{DatastreamDSWS2R-package}
\alias{DatastreamDSWS2R}
\title{DatastreamDSWS2R}
\description{
A package to manage access to the LSEG Datastream
 DSWS webservice
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CharlesCara/DatastreamDSWS2R}
  \item Report bugs at \url{https://github.com/CharlesCara/DatastreamDSWS2R/issues}
}

}
\author{
\strong{Maintainer}: Charles Cara \email{charles.cara@absolute-strategy.com}

Authors:
\itemize{
  \item Charles Cara \email{charles.cara@absolute-strategy.com}
}

}
