% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{dotconvert_JSON_Bool}
\alias{dotconvert_JSON_Bool}
\alias{.convert_JSON_Bool}
\title{.convert_JSON_Date}
\usage{
.convert_JSON_Bool(Input_Strings)
}
\arguments{
\item{Input_Strings}{a JSON string (or an array)}
}
\value{
an array of Dates
}
\description{
this is a modification of the function provided by
 'phiver' on
http://stackoverflow.com/questions/32076957/nas-introduced-when-
transforming-json-date
}
\details{
converts a JSON Boolean string to an R
 logical object
}
