% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{dotconvert_JSON_Datetime}
\alias{dotconvert_JSON_Datetime}
\alias{.convert_JSON_Datetime}
\title{.convert_JSON_Datetime}
\usage{
.convert_JSON_Datetime(Input_Strings)
}
\arguments{
\item{Input_Strings}{a JSON Date string (or an array)}
}
\value{
an array POSIXct/POSIXt object
}
\description{
this is a modification of the function provided by
'phiver' on
http://stackoverflow.com/questions/32076957/nas-introduced-when
-transforming-json-date
}
\details{
converts a JSON Date string (including with Timezone) to a
 POSIXct/POSIXt object
}
