% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{dotconvert_JSON_Double}
\alias{dotconvert_JSON_Double}
\alias{.convert_JSON_Double}
\title{.convert_JSON_Double}
\usage{
.convert_JSON_Double(Input_Strings)
}
\arguments{
\item{Input_Strings}{a JSON Date string (or an array)}
}
\value{
an array of Dates
}
\description{
this is a modification of the function provided by
 'phiver' on
http://stackoverflow.com/questions/32076957/nas-introduced-when-
transforming-json-date
}
\details{
converts a JSON Double object an R
 Double object
}
