% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{dotconvertJSONString}
\alias{dotconvertJSONString}
\alias{.convertJSONString}
\title{.convertJSONString}
\usage{
.convertJSONString(x)
}
\arguments{
\item{x}{the JSON string}
}
\value{
the parsed result: either Date, String or numeric
}
\description{
this is a modification of the function provided by
'phiver' on
http://stackoverflow.com/questions/32076957/nas-introduced-when
-transforming-json-date
}
\details{
converts a JSON string (including with Timezone) into either a
numeric, string, or a R Date object
}
\keyword{internal}
