% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCTSUpload.R
\name{dotEncryptPassword}
\alias{dotEncryptPassword}
\alias{.EncryptPassword}
\title{Encrypt the Datastream password}
\usage{
.EncryptPassword(strPassword = "")
}
\arguments{
\item{strPassword}{the password to be encrypted}
}
\value{
an encrypted password
}
\description{
This is a port of the VBA code
}
\keyword{internal}
