% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{dotgetSymbol}
\alias{dotgetSymbol}
\alias{.getSymbol}
\title{.getSymbol}
\usage{
.getSymbol(x)
}
\arguments{
\item{x}{a list that is expected to have an item 'Value'}
}
\value{
the parsed result: either Date, String or numeric
}
\description{
this is a modification of the function provided
by 'phiver' on
http://stackoverflow.com/questions/32076957/nas-introduced-when
-transforming-json-date
}
\details{
extracts and converts a JSON string (including with Timezone)
 from the item 'Symbol'
in the list x.  the item is parsed from JSON into either a numeric,
 string, or a R Date object
}
