% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCTSUpload.R
\name{dotgetTimeseries}
\alias{dotgetTimeseries}
\alias{.getTimeseries}
\title{convert xts timeseries into a string that can be sent to
the Datastream server}
\usage{
.getTimeseries(Data, freq, digits, NA_VALUE)
}
\arguments{
\item{Data}{the xts timeseries to be converted}

\item{freq}{the frequency of the data}

\item{digits}{the number of decimal places to round the data to}

\item{NA_VALUE}{the string to replace NA data with}
}
\value{
A string of the core data of Data
}
\description{
convert xts timeseries into a string that can be sent to
the Datastream server
}
\keyword{internal}
