% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConstructor.R
\docType{class}
\name{dsws-class}
\alias{dsws}
\alias{dsws-class}
\title{dsws}
\description{
An R5/RC object for accessing the Thomson Reuters Datastream DSWS service.
}
\details{
Creates an R5/RC4 object for accessing the Thomson Reuters Datastream DSWS service
}
\section{Methods}{

\describe{
\item{\code{basicRequest(instrument, datatype = "", expression = "", isList = FALSE,
  startDate, endDate, frequency = "D", kind = 0, format = "ByInstrument")}}{Return a timeSeriesRequest from Datastream dsws.  Should request either a datatype or an expression
not both.  If a datatype is provided then anythink in Expression will be ignored}

\item{\code{buildRequestList(frequency, instrument, datatype, expression, isList, startDate,
  endDate, kind, token)}}{Internal function that builds a request list that can be then parsed to JSON and sent to the
DSWS server}

\item{\code{getToken()}}{Returns a Token from the the dsws server that gives permission to access data}

\item{\code{initialize(dsws.serverURL = "", username = "", password = "",
  connect = TRUE)}}{initialises the class and unless noConnect is TRUE connects to the Datastream dsws server }

\item{\code{listRequest(instrument, datatype = "", expression = "", requestDate)}}{Return a listRequest from Datastream dsws.  instrument should contain a list mnemonic,
such as 'LFTSE100'  }

\item{\code{makeRequest()}}{make a request from the DSWS server.  The request (in a R list form) is taken from
.self$requestList, parsed into JSON and sent to the DSWS server.  The JSON response
is parsed and saved in .self$dataResponse}

\item{\code{snapshotRequest(instrument, datatype = "", expression = "", requestDate)}}{Return a snapshot (static) Request from Datastream dsws.
instrument should contain a constituent list eg "LFTSE100"}

\item{\code{timeSeriesListRequest(instrument, datatype = "", expression = "", startDate,
  endDate, frequency = "D", format = c("ByInstrument", "ByDatatype"))}}{Return a timeSeriesListRequest from Datastream dsws}

\item{\code{timeSeriesRequest(instrument, datatype = "", expression = "", startDate,
  endDate, frequency = "D", format = c("ByInstrument", "ByDatatype"))}}{Return a timeSeriesRequest from Datastream dsws.  Should request either a datatype or an expression
not both.  If a datatype is provided then anythink in Expression will be ignored}
}}

