% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{getDataStream}
\alias{getDataStream}
\title{Initialise connection with Datastream DSWS server}
\usage{
getDataStream(dweURLwsdl = "", User = as.character("USERNAME"),
  Pass = as.character("PASSWORD"))
}
\arguments{
\item{dweURLwsdl}{The URL of the server}

\item{User}{The username for Datastream.  If not provided it will use the username from the windows registry}

\item{Pass}{The password for Datastream.  Also sourced from Registry}
}
\value{
a DWE connection object
}
\description{
\code{getDataStream} initialises the connection with the Datastream DWE server
}
\details{
Initialise connection with Datastream DSWS server
}

