% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{listRequest}
\alias{listRequest}
\title{Make a list request for static data (Depreciated)}
\usage{
listRequest(
  dwei = getDataStream(),
  DSCode,
  Expression = "",
  startDate = Sys.Date(),
  endDate = Sys.Date(),
  frequency = "D",
  verbose = FALSE
)
}
\arguments{
\item{dwei}{- A Datastream Client Interface object created with
getDataStream}

\item{DSCode}{- the constituent list for the request eg LDJSTOXX}

\item{Expression}{- the data to return eg MNEM or NAME.
If NULL or "" then we will return the code that has been loaded
into the User Created List.}

\item{startDate}{- the date of the request, or the string "TODAY"}

\item{endDate}{- Ignored}

\item{frequency}{- the frequency of the request}

\item{verbose}{- whether to give messages during the request}
}
\value{
returns an array of the requested information
}
\description{
\code{listRequest} Function that returns a the value of
Expression for the instrument list in DSCode from Datastream
}
\details{
Make a list request for static data
}
