% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{myStaticRequestSet}
\alias{myStaticRequestSet}
\title{myStaticRequestSet}
\usage{
myStaticRequestSet(mydsws = dsws$new(), instrument, iExpression,
  endDate = Sys.Date(), frequency = "D")
}
\arguments{
\item{mydsws}{a dsws object, if not provided a new one will be created}

\item{instrument}{array of instruments}

\item{iExpression}{an expression such as PCH#(XXXX,1M)}

\item{endDate}{the date of the request}

\item{frequency}{optional frequency defaults to "D"}
}
\value{
a dataframe of the
}
\description{
internal function for requesting an expression for an array of instruments.
The function will initially try a snapshot request, and if this
fails try a timeseries request.
}
