% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCTSUpload.R
\name{processUpload}
\alias{processUpload}
\title{Process a user object which might be a timeseries or an index}
\usage{
processUpload(UserObject, UserObjectType, URL, mydsws)
}
\description{
Process a user object which might be a timeseries or an index
}
