% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{staticListRequestSet}
\alias{staticListRequestSet}
\title{staticListRequestSet}
\usage{
staticListRequestSet(mydsws = dsws$new(), instrument, expression = "",
  endDate = Sys.Date(), frequency = "D")
}
\arguments{
\item{mydsws}{a dsws object, if not provided a new one will be created}

\item{instrument}{array of instruments}

\item{expression}{an array of expressions such as PCH#(XXXX,1M)}

\item{endDate}{the date of the request}

\item{frequency}{optional frequency defaults to "D"}
}
\value{
a dataframe of the data
}
\description{
This function creates a dataframe set of static list requests for a constituent list
}
\details{
This function creates a dataframe set of static list requests for a constituent list
}
