% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{staticRequest}
\alias{staticRequest}
\title{make a static request}
\usage{
staticRequest(dwei = getDataStream(), DSCode, Expression = "",
  endDate = Sys.Date(), frequency = "D", verbose = FALSE,
  noCache = FALSE)
}
\arguments{
\item{dwei}{- A Datastream Client Interface object created with getDataStream}

\item{DSCode}{- an array of instruments eg c("RIO","MKS")}

\item{Expression}{- the data to return eg MNEM or NAME}

\item{endDate}{- the date of the request, or the string "TODAY"}

\item{frequency}{- the frequency of the request}

\item{verbose}{- whether to give messages during the request}
}
\value{
returns an array of the requested information
}
\details{
\code{staticRequest} Function that returns a the value of Expression for the array of instruments in DSCode from Datastream
parameters are
}

