% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{staticRequestSet}
\alias{staticRequestSet}
\title{staticRequestSet}
\usage{
staticRequestSet(
  mydsws = dsws$new(),
  instrument,
  expression = "",
  endDate = Sys.Date(),
  frequency = "D",
  verbose = FALSE
)
}
\arguments{
\item{mydsws}{a dsws object, if not provided a new one will be created}

\item{instrument}{array of instruments}

\item{expression}{an array of expressions such as PCH#(XXXX,1M) or Dataitems}

\item{endDate}{the date of the request}

\item{frequency}{optional frequency defaults to "D"}

\item{verbose}{whether to display messages as making the request}
}
\value{
a dataframe of the data
}
\description{
This function creates a dataframe set of static requests for a set of stocks/indices
}
\details{
return a dataframe of static data
}
