% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{timeSeriesListRequest}
\alias{timeSeriesListRequest}
\title{make a timeSeries request for a list (Depreciated)}
\usage{
timeSeriesListRequest(dwei = getDataStream(), DSCode, Instrument, startDate,
  endDate = Sys.Date(), frequency = "D", sStockList, aTimeSeries,
  verbose = FALSE)
}
\arguments{
\item{dwei}{- A Datastream Client Interface object created with
getDataStream}

\item{DSCode}{- the constituent list requested eg 'LFTSE100'}

\item{Instrument}{- the expression to return for each member
of constituent list}

\item{startDate}{- the start date of the timeseries}

\item{endDate}{- the end date of the timeseries}

\item{frequency}{- the frequency of the request}

\item{sStockList}{- variable that is returned with list of of the stocks}

\item{aTimeSeries}{- variable that is returned with the set of timeseries}

\item{verbose}{- whether to give messages during the request}
}
\value{
whether the request has been successful
   , but also
   in sStockList: a list a two element vector of the displayname
    and symbol for each timeseries
   in aTimeseries: a list of class xts with the requested
    timeseries information
}
\description{
make a timeseries request for a constituent list from
 Datastream DSWS
\code{timeSeriesListRequest} Function that returns a timeseries from
Datastream constituent list parameters are
}
\details{
List request
}
