% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{timeSeriesRequest}
\alias{timeSeriesRequest}
\title{make a timeseries request}
\usage{
timeSeriesRequest(dwei = getDataStream(), DSCodes = "", Instrument = "",
  startDate = Sys.Date(), endDate = Sys.Date(), frequency = "D",
  sStockList, aTimeSeries, myType = "numeric", verbose = FALSE)
}
\arguments{
\item{dwei}{- A Datastream Client Interface object created with getDataStream}

\item{DSCodes}{- one or more codes to return, eg "MKS" or c("MKS","SAB")}

\item{Instrument}{- the instrument or expression to return eg PCH#(XXXX,1M)}

\item{startDate}{- the start date of the timeseries}

\item{endDate}{- the end date of the timeseries}

\item{frequency}{- the frequency of the request}

\item{sStockList}{- variable that is returned with list of of the stocks}

\item{aTimeSeries}{- variable that is returned with the set of timeseries.  This is a list that is not
guaranteed to be in the same order as sStockList}

\item{myType}{- the type of the return values eg numeric (default), Date or Character}

\item{verbose}{- whether to give messages during the request}
}
\value{
whether the request has been successful
   in sStockList: a list a two element vector of the displayname and symbol for each timeseries
   in aTimeseries: a list of class xts with the requested timeseries information
}
\description{
make a timeseries request from the Datastream DSWS server
}
\details{
\code{timeSeriesRequest} Function that obtains a timeseries from Datastream
parameters are
}

