% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{currencyDS2ISO}
\alias{currencyDS2ISO}
\title{Conversion table of Datastream to ISO currency codes}
\format{
A data frame with 161 rows and 3 variables:
\describe{
  \item{dsCode}{the datastream code}
  \item{isoCode}{the ISO code for the currency}
  \item{primeCode}{primaryCode for currency or alternative}
  \item{Multiplier}{the units of the currency}
}
}
\usage{
currencyDS2ISO
}
\description{
Conversion table of Datastream to ISO currency codes
}
\keyword{datasets}
