% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{getDataStream}
\alias{getDataStream}
\title{Initialise connection with Datastream DSWS server}
\usage{
getDataStream(dweURLwsdl = "", User = as.character("USERNAME"),
  Pass = as.character("PASSWORD"))
}
\arguments{
\item{dweURLwsdl}{Ignored}

\item{User}{Ignored - now sourced from options()$Datastream.Username}

\item{Pass}{Ignored - now sourced from options()$Datastream.Password}
}
\value{
a dsws object
}
\description{
\code{getDataStream} initialises an R5 object that contains
 a connection with the Datastream DWE server.  This function has been
  provided for backward compatibility
}
\details{
Initialise connection with Datastream DSWS server.  Provided
for backwards compatibility
}
